Only include JS when the user is logged in
This ensures that 'sensitive' CMS edit links aren't leaked to all users, and the JS is only included when it's needed (e.g. when a user is logged in).

This might not be a good idea if you want the edit link to always work all the time, but prevents a pretty minor security risk (exposing the CMS edit link to attackers), and also helps performance (not including an additional HTTP request to the server that normal website users don't need).

The reason it's only a pretty minor security risk is that the CMS edit link is already exposed if you know the website runs SilverStripe, so this PR only helps with a bit of obscurity.